+++ CHANGELOG.md
@@ -4,18 +4,45 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.23] (2023-4-12)
+[0.23]: https://github.com/CoboCustody/cobo-python-api/compare/0.22...0.23
+### Added
+- Add Get standard tx by request_ids. https://github.com/CoboCustody/cobo-python-api/pull/36
+
+## [0.22] (2023-4-10)
+[0.22]: https://github.com/CoboCustody/cobo-python-api/compare/0.21...0.22
+### Changed
+- Update request nonce. https://github.com/CoboCustody/cobo-python-api/pull/37
+
+## [0.21] (2023-3-13)
+[0.21]: https://github.com/CoboCustody/cobo-python-api/compare/0.20...0.21
+### Changed
+- Update MPC wallet client. https://github.com/CoboCustody/cobo-python-api/pull/30, https://github.com/CoboCustody/cobo-python-api/pull/31
+
+## [0.20] (2023-2-16)
+[0.20]: https://github.com/CoboCustody/cobo-python-api/compare/0.19...0.20
+### Changed
+- Update MPC wallet client. https://github.com/CoboCustody/cobo-python-api/pull/25, https://github.com/CoboCustody/cobo-python-api/pull/27
+- Update Web3 wallet client. https://github.com/CoboCustody/cobo-python-api/pull/28
+
+## [0.19] (2022-12-22)
+[0.19]: https://github.com/CoboCustody/cobo-python-api/compare/0.18...0.19
+
+### Changed
+- Update MPC wallet client. https://github.com/CoboCustody/cobo-python-api/pull/22, https://github.com/CoboCustody/cobo-python-api/pull/23
+
 ## [0.18] (2022-12-18)
 [0.18]: https://github.com/CoboCustody/cobo-python-api/compare/0.17...0.18
 
 ### Changed
-- Update MPC wallet Saas api.https://github.com/CoboCustody/cobo-python-api/pull/17, https://github.com/CoboCustody/cobo-python-api/pull/18, https://github.com/CoboCustody/cobo-python-api/pull/20, https://github.com/CoboCustody/cobo-python-api/pull/21
+- Update MPC wallet SaaS api.https://github.com/CoboCustody/cobo-python-api/pull/17, https://github.com/CoboCustody/cobo-python-api/pull/18, https://github.com/CoboCustody/cobo-python-api/pull/20, https://github.com/CoboCustody/cobo-python-api/pull/21
 
 
 ## [0.17] (2022-12-02)
 [0.17]: https://github.com/CoboCustody/cobo-python-api/compare/0.16...0.17
 
 ### Added
-- Support MPC wallet Saas api.https://github.com/CoboCustody/cobo-python-api/pull/12, https://github.com/CoboCustody/cobo-python-api/pull/13, https://github.com/CoboCustody/cobo-python-api/pull/14, https://github.com/CoboCustody/cobo-python-api/pull/15, https://github.com/CoboCustody/cobo-python-api/pull/16
+- Support MPC wallet SaaS api.https://github.com/CoboCustody/cobo-python-api/pull/12, https://github.com/CoboCustody/cobo-python-api/pull/13, https://github.com/CoboCustody/cobo-python-api/pull/14, https://github.com/CoboCustody/cobo-python-api/pull/15, https://github.com/CoboCustody/cobo-python-api/pull/16
 
 ## [0.16] (2022-11-14)
 [0.16]: https://github.com/CoboCustody/cobo-python-api/compare/0.15...0.16
@@ -27,7 +54,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 [0.15]: https://github.com/CoboCustody/cobo-python-api/compare/0.14...0.15
 
 ### Added
-- Support Web3 wallet Saas api. https://github.com/CoboCustody/cobo-python-api/pull/10, https://github.com/CoboCustody/cobo-python-api/pull/11
+- Support Web3 wallet SaaS api. https://github.com/CoboCustody/cobo-python-api/pull/10, https://github.com/CoboCustody/cobo-python-api/pull/11
 
 
 ## [0.13] (2022-07-12)